Give the two graphs distinctive names
One is a graph of instance groups and the tracking data for what
capacity they have and what we've commited to in the run of the task
manager. The other is a DependencyGraph for tracking job dependencies.

Also, use the same method to zero out the graph in the two places we
might want to do it. The main reason we would ever actually zero out the
graph in awx/main/managers.py is in tests where we have not already
initialized the instance group capacity graph.